server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
openai==1.3.4
google-cloud-texttospeech==2.15.1
langchain==0.0.335
langgraph==0.0.35
langchain-openai==0.0.5
python-dotenv==1.0.0
asyncio==3.4.3
aiofiles==23.2.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
cachetools==5.3.2
tenacity==8.2.3
loguru==0.7.2